permissions

+++ app/controllers/v1/forms/config/mistake_types_controller.rb
@@ -1,20 +1,21 @@
 class V1::Forms::Config::MistakeTypesController < ApplicationController
   before_action :authenticate_v1_user!
-  respond_to :json
 
   def index
+    authorize ::Forms::Config::MistakeType
     types = ::Forms::Config::MistakeType.order(name: :asc).page(params[:page]).per(params[:per])
     render json: types, meta: pagination_dict(types)
   end
 
   def show
     type = ::Forms::Config::MistakeType.find params[:id]
+    authorize type
     render json: type
   end
 
   def update
     type = ::Forms::Config::MistakeType.find params[:id]
-
+    authorize type
     if type.update_attributes(mistake_type_params)
       render json: type
     else
@@ -24,6 +25,7 @@ class V1::Forms::Config::MistakeTypesController < ApplicationController
   end
 
   def create
+    authorize ::Forms::Config::MistakeType
     type = ::Forms::Config::MistakeType.new mistake_type_params
     if type.save
       render json: type
@@ -34,6 +36,7 @@ class V1::Forms::Config::MistakeTypesController < ApplicationController
 
   def destroy
     type = ::Forms::Config::MistakeType.find params[:id]
+    authorize type
     type.destroy
     render json: {}
   end
